cleanup and simplify module
The main change is that I made FromLucid not be passed getCurrentRoute.
Most of the time, that is not needed, and if a Handler does need it, it can
query yesod for that itself, before calling the lucid function.

Also, added the type LucidHtml = Html ()
The motivation for this is that both yesod and lucid export different Html
types, and this allow easily marking Handler LucidHtml in a user's code.